cleanup(event): move all formatter code to types::format
