+++ falco_event/src/types/string/cstr.rs
@@ -1,7 +1,5 @@
 use crate::fields::{FromBytes, FromBytesError, ToBytes};
-use crate::types::ByteBufFormatter;
 use std::ffi::CStr;
-use std::fmt::{Debug, Formatter};
 use std::io::Write;
 
 impl<'a> FromBytes<'a> for &'a CStr {
@@ -28,18 +26,6 @@ impl ToBytes for &CStr {
     #[inline]
     fn default_repr() -> impl ToBytes {
         0u8
-    }
-}
-
-/// Falco-style CStr formatter
-///
-/// Formats the string like a byte buffer (replacing non-ASCII characters with `.`).
-/// See [`ByteBufFormatter`] for the implementation.
-pub struct CStrFormatter<'a>(pub &'a CStr);
-
-impl Debug for CStrFormatter<'_> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        Debug::fmt(&ByteBufFormatter(self.0.to_bytes()), f)
     }
 }
