+++ falco_event/src/types/string/mod.rs
@@ -2,6 +2,5 @@ mod cstr;
 mod cstr_array;
 mod cstr_pair_array;
 
-pub use cstr::CStrFormatter;
 pub use cstr_array::CStrArray;
 pub use cstr_pair_array::CStrPairArray;
