+++ falco_schema_derive/src/format.rs
@@ -6,8 +6,8 @@ pub fn display_wrapper_for(
     val_tt: proc_macro2::TokenStream,
 ) -> proc_macro2::TokenStream {
     match pt_type.to_string().as_str() {
-        "PT_BYTEBUF" => quote!(#val_tt.map(|t| crate::format::ByteBufFormatter(t))),
-        "PT_CHARBUF" => quote!(#val_tt.map(|t| crate::format::CStrFormatter(t))),
+        "PT_BYTEBUF" => quote!(#val_tt.map(|t| crate::types::format::ByteBufFormatter(t))),
+        "PT_CHARBUF" => quote!(#val_tt.map(|t| crate::types::format::CStrFormatter(t))),
         "PT_FSPATH" => quote!(#val_tt.map(|p| p.display())),
         _ => val_tt,
     }
